fix: export locking state

+++ app/schedule/export_archive_data.js
@@ -6,9 +6,9 @@ const Subscription = require('egg').Subscription;
 class exportArchiveData extends Subscription {
 
   static get schedule() {
-    // export archive data every 5s
+    // export archive data every 4s
     return {
-      interval: '2s',
+      interval: '4s',
       type: 'worker',
     };
   }
@@ -16,8 +16,13 @@ class exportArchiveData extends Subscription {
   async subscribe() {
     this.excludeAttributes = this.app.config.exportExcludeAttributes;
     if (!this.app.config.exportArchiveBaseDir) return;
-    rimraf.sync(this.app.config.exportArchiveBaseDir);
-    await this.ctx.service.database.exportData();
+    if (this.app.EXPORT_LOCKING === true) return;
+    try {
+      this.app.EXPORT_LOCKING = true;
+      rimraf.sync(this.app.config.exportArchiveBaseDir);
+      await this.ctx.service.database.exportData();
+    } catch (_) { /* ignore */ }
+    this.app.EXPORT_LOCKING = false;
   }
 }
